# Faster Integer.sqrt for large bignum